refactor gender tests
